[#109] implemented initial BaseXWorkload & SteeringMixerWorkload
