+++ knitware-bots/barr-e/spine/main/CMakeLists.txt
@@ -1,7 +1,7 @@
 # knitware-bots/barr-e/spine/main/CMakeLists.txt
 
 idf_component_register(
-    SRCS "main.cpp"
+    SRCS "spine_main.cpp"
     INCLUDE_DIRS "."
     PRIV_REQUIRES robotick-engine
 )
