+++ knitware-bots/barr-e/spine/main/spine_main.cpp
@@ -1,6 +1,8 @@
 // Copyright Robotick Labs
 // SPDX-License-Identifier: Apache-2.0
 
+#include "../../shared/model.h"
+
 #include "robotick/framework/Engine.h"
 #include "robotick/framework/Model.h"
 #include "robotick/framework/registry/WorkloadRegistry.h"
@@ -22,31 +24,22 @@ namespace robotick
 {
 	void ensure_workloads()
 	{
-		// ROBOTICK_KEEP_WORKLOAD(ConsoleTelemetryWorkload)
+		ROBOTICK_KEEP_WORKLOAD(SteeringMixerWorkload)
+		ROBOTICK_KEEP_WORKLOAD(BaseXWorkload)
 		ROBOTICK_KEEP_WORKLOAD(HeartbeatDisplayWorkload)
-		// ROBOTICK_KEEP_WORKLOAD(TimingDiagnosticsWorkload)
-		// ROBOTICK_KEEP_WORKLOAD(SyncedGroupWorkload)
+		ROBOTICK_KEEP_WORKLOAD(SequencedGroupWorkload)
 	}
 
 } // namespace robotick
-
-void create_simple_model(robotick::Model& model)
-{
-	auto root = model.add("HeartbeatDisplayWorkload", "heartbeat", 30.0);
-	model.set_root(root);
-}
 
 void run_engine_on_core1(void* param)
 {
 	ESP_LOGI(ENGINE_TASK_NAME, "Running on CPU%d", xPortGetCoreID());
 
-	esp_task_wdt_add(nullptr); // register current task with watchdog
-
 	auto* engine = static_cast<robotick::Engine*>(param);
 
 	robotick::Model model;
-	// create_threaded_model(model);
-	create_simple_model(model);
+	barr_e::populate_model_spine(model);
 
 	ESP_LOGI("Robotick", "Loading Robotick model...");
 	engine->load(model); // Ensures memory locality on Core 1
@@ -61,6 +54,13 @@ void run_engine_on_core1(void* param)
 extern "C" void app_main(void)
 {
 	ESP_LOGI("main_task", "Started on CPU%d", xPortGetCoreID());
+    
+	// Kill Watchdog for this task (your main thread)
+    esp_task_wdt_delete(nullptr);
+
+    // Disable the watchdog for the idle task on Core 1
+    TaskHandle_t idle1 = xTaskGetIdleTaskHandleForCore(1);
+    esp_task_wdt_delete(idle1);
 
 	robotick::ensure_workloads();
